flask demo: don't provide a meaningless default server
Instead, ask the user to provide the test server. Which is annoying,
but...

The flask demo needs an OAuth-capable server, and we are currently
using a no-auth server, so the user just sees a message "nothing to
demo here".

This is because:
- The old server that presumably worked was taken down.
- We put in the new R4 modern test server.
- The modern server has a temporary bug where its metadata doesn't
  validate
- So we didn't notice that the flask server wouldn't work even if
  validation passed.

All of the public test servers I know about require you to email
someone to set up a user. So it's not obvious what a good replacement
would be.

Honestly, we should probably just add a new default demo that supports
no-auth servers and still shows something interesting.

But for now, this gets the flask demo to a more honest place and asks
the user to provide their own server. :shrug: